Adding integration into configuration file and interactive mode. Integrating Pinvoke calls into already existing Advapi32 and Kernel32 classes
